API: Disallow show_traces=True with time_viewer=False [circle full]
